Move all file generation logic to new FileGenerator class, Add unit test for header method, header method update to validate struct file name is Capitalized.
