+++ Sources/Chroma/App/Chroma.swift
@@ -7,7 +7,6 @@
 //
 
 import ArgumentParser
-import Files
 import Foundation
 
 public struct Chroma: ParsableCommand {
@@ -26,42 +25,19 @@ public struct Chroma: ParsableCommand {
     public init() {}
     
     public func run() throws {
-        let outputFile = try createOutputFile()
-        let content = try getContentFromAssetsFile(outputFile: outputFile)
-        try outputFile.write(content)
+        let generator = FileGenerator(
+            asset: asset,
+            path: path,
+            type: type,
+            framework: framework
+        )
+        let file = try generator.generate()
         print(
             """
-            \(outputFile.name) was generated successfully.
-            Can be found at \(outputFile.path)
+            \(file.name) was generated successfully.
+            Can be found at \(file.path)
             """
         )
     }
     
-}
-
-extension Chroma {
-    private func createOutputFile() throws -> File {
-        // Check if path param is a valid swift file path
-        guard let pathURL = URL(string: path), !pathURL.hasDirectoryPath, pathURL.pathExtension == "swift"  else {
-            throw ChromaError.invalidPath(path: path)
-        }
-        
-        let folder = try Folder(path: pathURL.deletingLastPathComponent().path)
-        return try File(named: pathURL.lastPathComponent, at: folder)
-    }
-    
-    private func getContentFromAssetsFile(outputFile: File) throws -> String {
-        let assetFolder = try Folder(path: asset)
-        let body = framework.fileBody(asset: assetFolder).joined(separator: "\n")
-        return framework.fileContent(header: header(file: outputFile), body: body)
-    }
-    
-    private func header(file: File) -> String {
-        switch type {
-        case .extension:
-            return "\(type.rawValue) \(framework.variableType)"
-        case .struct:
-            return "\(type.rawValue) \(file.nameExcludingExtension)"
-        }
-    }
 }
