+++ Sources/Chroma/App/Framework.swift
@@ -7,7 +7,6 @@
 
 import Foundation
 import ArgumentParser
-import Files
 
 enum Framework: String, ExpressibleByArgument {
     case AppKit
@@ -16,8 +15,6 @@ enum Framework: String, ExpressibleByArgument {
 }
 
 extension Framework {
-    private static let colorAssetExtension = "colorset"
-    
     var defaultValue: String {
         switch self {
         case .UIKit, .AppKit:
@@ -51,55 +48,6 @@ extension Framework {
         case .SwiftUI:
             return ["accentColor"]
         }
-    }
-    
-    func fileContent(header: String, body: String) -> String {
-        """
-        //
-        //  Generated by Chroma.
-        //  https://github.com/jjotaum/Chroma.
-        //
-        //  This file was auto generated please do not modify it directly.
-        //
-        
-        import \(rawValue)
-        
-        \(header) {
-        
-        \(body)
-        
-        }
-        """
-    }
-    
-    func fileBody(asset: Folder) -> Array<String> {
-        let assetKey = asset.nameExcludingExtension
-        // Get subfolders with valid extension
-        let colorSubfolders = asset.subfolders.recursive.filter { $0.extension == Self.colorAssetExtension }
-        // Group them by parent name to use them as MARK's
-        let parents = Dictionary(grouping: colorSubfolders, by: \.parent?.nameExcludingExtension)
-        var content: [String] = []
-        // Add main variables
-        if let root = parents[assetKey] {
-            content.append(contentsOf: colorVariableNames(folders: root))
-        }
-        // Sort keys to give output MARKS alphabetical order
-        parents.sorted(by: { $0.key ?? "" < $1.key ?? "" }).forEach { (key, value) in
-            guard key != assetKey else { return }
-            if let mark = key {
-                content.append("    // MARK: - \(mark)")
-            }
-            // Add MARK variables
-            content.append(contentsOf: colorVariableNames(folders: value))
-        }
-        return content
-    }
-    
-    private func colorVariableNames(folders: [Folder]) -> [String] {
-        // We filter out duplicated variable names
-        Set(folders.compactMap { colorFolder in
-            return colorVariable(name: colorFolder.nameExcludingExtension)
-        }).sorted()
     }
     
     func colorVariable(name: String) -> String? {
